Changed clients integration test to category unit and moved it to a new file and directory.
