+++ src/LodeRunner.API.Test/IntegrationTests/Controllers/Clients.cs
@@ -50,35 +50,6 @@ namespace LodeRunner.API.Test.IntegrationTests.Controllers
         }
 
         /// <summary>
-        /// Gets endpoints return success and correct content.
-        /// </summary>
-        /// <param name="url">The URL.</param>
-        /// <param name="expectedValue">The expectedValue.</param>
-        /// <returns><see cref="Task"/> representing the asynchronous unit test.</returns>
-        [Theory]
-        [Trait("Category", "Integration")]
-        [InlineData("/", "text/html; charset=utf-8")]
-        [InlineData("/version", "text/plain")]
-        public async Task CanGetEndpointsReturnSuccessAndCorrectContentType(string url, string expectedValue)
-        {
-            using var httpClient = ComponentsFactory.CreateLodeRunnerAPIHttpClient(this.factory);
-
-            var httpResponse = await httpClient.GetAsync(url);
-
-            if (httpResponse.StatusCode == HttpStatusCode.OK)
-            {
-                this.output.WriteLine($"UTC Time:{DateTime.UtcNow}\tAction: GET Url '{url}'\tResponse StatusCode: 'OK'");
-            }
-            else
-            {
-                this.output.WriteLine($"UTC Time:{DateTime.UtcNow}\tAction: GET Url '{url}'\tUNEXPECTED Response StatusCode: '{httpResponse.StatusCode}'");
-            }
-
-            var responseContents = await httpResponse.Content.ReadAsStringAsync();
-            AssertExtension.Equal(expectedValue, httpResponse.Content.Headers.ContentType.ToString(), responseContents);
-        }
-
-        /// <summary>
         /// Determines whether this instance [can get clients].
         /// </summary>
         /// <returns><see cref="Task"/> representing the asynchronous unit test.</returns>
