+++ src/LodeRunner.API.Test/UnitTests/Controllers.cs
@@ -1,0 +1,82 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Net.Http;
+using System.Net.Http.Json;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using System.Threading.Tasks;
+using LodeRunner.API.Models;
+using LodeRunner.API.Test.IntegrationTests;
+using LodeRunner.API.Test.IntegrationTests.Extensions;
+using LodeRunner.Core.Models;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace LodeRunner.API.Test.UnitTests
+{
+    /// <summary>
+    /// Represents Controllers.
+    /// </summary>
+    public class Controllers : IClassFixture<ApiWebApplicationFactory<Startup>>
+    {
+        private readonly ApiWebApplicationFactory<Startup> factory;
+
+        private readonly JsonSerializerOptions jsonOptions;
+
+        private readonly ITestOutputHelper output;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Controllers"/> class.
+        /// </summary>
+        /// <param name="factory">The factory.</param>
+        /// <param name="output">The output.</param>
+        public Controllers(ApiWebApplicationFactory<Startup> factory, ITestOutputHelper output)
+        {
+            this.factory = factory;
+
+            this.output = output;
+
+            this.jsonOptions = new()
+            {
+                DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
+                PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+                DictionaryKeyPolicy = JsonNamingPolicy.CamelCase,
+                PropertyNameCaseInsensitive = true,
+            };
+            this.jsonOptions.Converters.Add(new JsonStringEnumConverter());
+        }
+        
+        /// <summary>
+        /// Gets endpoints return success and correct content.
+        /// </summary>
+        /// <param name="url">The URL.</param>
+        /// <param name="expectedValue">The expectedValue.</param>
+        /// <returns><see cref="Task"/> representing the asynchronous unit test.</returns>
+        [Theory]
+        [Trait("Category", "Unit")]
+        [InlineData("/", "text/html; charset=utf-8")]
+        [InlineData("/version", "text/plain")]
+        public async Task CanGetEndpointsReturnSuccessAndCorrectContentType(string url, string expectedValue)
+        {
+            using var httpClient = ComponentsFactory.CreateLodeRunnerAPIHttpClient(this.factory);
+
+            var httpResponse = await httpClient.GetAsync(url);
+
+            if (httpResponse.StatusCode == HttpStatusCode.OK)
+            {
+                this.output.WriteLine($"UTC Time:{DateTime.UtcNow}\tAction: GET Url '{url}'\tResponse StatusCode: 'OK'");
+            }
+            else
+            {
+                this.output.WriteLine($"UTC Time:{DateTime.UtcNow}\tAction: GET Url '{url}'\tUNEXPECTED Response StatusCode: '{httpResponse.StatusCode}'");
+            }
+
+            var responseContents = await httpResponse.Content.ReadAsStringAsync();
+            AssertExtension.Equal(expectedValue, httpResponse.Content.Headers.ContentType.ToString(), responseContents);
+        }
+    }
+}
